[ch62729] fix regression + refacto for easier unit tests
