+++ custom-recipes/timeseries-preparation-extrema-extraction/recipe.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 import logging
+
 from dataiku.customrecipe import get_recipe_config
+
 from dku_timeseries import ExtremaExtractor
-from commons import check_python_version, get_input_output, get_extrema_extraction_params
-from recipe_config_loading import check_time_column_parameter, check_and_get_groupby_columns
+from io_utils import get_input_output
+from recipe_config_loading import check_time_column_parameter, check_and_get_groupby_columns, check_python_version, get_extrema_extraction_params
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, format='timeseries-preparation plugin %(levelname)s - %(message)s')
