+++ python-lib/dku_timeseries/__init__.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from dku_timeseries.resampling.resampling import ResamplerParams, Resampler
-from dku_timeseries.windowing import WindowAggregatorParams, WindowAggregator
-from dku_timeseries.interval_restriction import IntervalRestrictorParams, IntervalRestrictor
 from dku_timeseries.extrema_extraction import ExtremaExtractorParams, ExtremaExtractor
+from dku_timeseries.interval_restriction import IntervalRestrictorParams, IntervalRestrictor
+from dku_timeseries.resampling import ResamplerParams, Resampler
+from dku_timeseries.windowing import WindowAggregatorParams, WindowAggregator
