+++ tests/python/unit/dku_timeseries/resampling/test_resampler_helpers.py
@@ -2,8 +2,8 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from dku_timeseries.resampling.resampling import ResamplerParams
 from dku_timeseries.timeseries_helpers import generate_date_range, get_date_offset
+from recipe_config_loading import get_resampling_params
 
 
 @pytest.fixture
@@ -12,26 +12,6 @@ def config():
               u'datetime_column': u'Date', u'advanced_activated': False, u'time_unit': u'quarters', u'clip_start': 0, u'time_step': 2,
               u'interpolation_method': u'linear'}
     return config
-
-
-def get_params(config):
-    def _p(param_name, default=None):
-        return config.get(param_name, default)
-
-    interpolation_method = _p('interpolation_method')
-    extrapolation_method = _p('extrapolation_method')
-    constant_value = _p('constant_value')
-    time_step = _p('time_step')
-    time_unit = _p('time_unit')
-    time_unit_end_of_week = _p('time_unit_end_of_week')
-
-    params = ResamplerParams(interpolation_method=interpolation_method,
-                             extrapolation_method=extrapolation_method,
-                             constant_value=constant_value,
-                             time_step=time_step,
-                             time_unit=time_unit,
-                             time_unit_end_of_week=time_unit_end_of_week)
-    return params
 
 
 class TestResamplerHelpers:
@@ -66,7 +46,7 @@ class TestResamplerHelpers:
 
     def test_generate_date_range_month(self, config):
         config["time_unit"] = "months"
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
@@ -98,7 +78,7 @@ class TestResamplerHelpers:
 
     def test_generate_date_range_week(self, config):
         config["time_unit"] = "weeks"
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
@@ -118,7 +98,7 @@ class TestResamplerHelpers:
 
         config["time_unit"] = "weeks"
         config["time_unit_end_of_week"] = "WED"
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
@@ -131,7 +111,7 @@ class TestResamplerHelpers:
         start_time = pd.Timestamp('2020-01-23 00:00:00')
         end_time = pd.Timestamp('2021-01-18 00:00:00')
 
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
@@ -145,7 +125,7 @@ class TestResamplerHelpers:
         start_time = pd.Timestamp('2020-01-01 00:00:00')
         end_time = pd.Timestamp('2021-06-18 00:00:00')
 
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
@@ -159,7 +139,7 @@ class TestResamplerHelpers:
         start_time = pd.Timestamp('2021-01-02 00:00:00')
         end_time = pd.Timestamp('2021-01-10 00:00:00')
 
-        params = get_params(config)
+        params = get_resampling_params(config)
         frequency = params.resampling_step
         time_unit = params.time_unit
         time_step = params.time_step
